classes from Request component folder moved to HttpFoundation folder
- HttpFoundation and Request is solving the same concept of problems
