+++ packages/framework/src/Component/HttpFoundation/DenyScriptNameInRequestPathListener.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Shopsys\FrameworkBundle\Component\Request;
+namespace Shopsys\FrameworkBundle\Component\HttpFoundation;
 
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
 use Symfony\Component\HttpKernel\Event\GetResponseEvent;
