+++ packages/framework/src/Component/HttpFoundation/VaryResponseByXRequestedWithHeaderListener.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Shopsys\FrameworkBundle\Component\Request;
+namespace Shopsys\FrameworkBundle\Component\HttpFoundation;
 
 use Symfony\Component\HttpKernel\Event\FilterResponseEvent;
